+++ src/main/java/org/betonquest/betonquest/BetonQuest.java
@@ -41,6 +41,8 @@ import org.betonquest.betonquest.database.GlobalData;
 import org.betonquest.betonquest.database.MySQL;
 import org.betonquest.betonquest.database.SQLite;
 import org.betonquest.betonquest.database.Saver;
+import org.betonquest.betonquest.feature.registry.CoreFeatureFactories;
+import org.betonquest.betonquest.feature.registry.FeatureRegistries;
 import org.betonquest.betonquest.item.QuestItemHandler;
 import org.betonquest.betonquest.listener.CustomDropListener;
 import org.betonquest.betonquest.listener.JoinQuitListener;
@@ -57,8 +59,6 @@ import org.betonquest.betonquest.playerhider.PlayerHider;
 import org.betonquest.betonquest.quest.registry.CoreQuestTypes;
 import org.betonquest.betonquest.quest.registry.QuestRegistry;
 import org.betonquest.betonquest.quest.registry.QuestTypeRegistries;
-import org.betonquest.betonquest.quest.registry.feature.CoreFeatureFactories;
-import org.betonquest.betonquest.quest.registry.feature.FeatureRegistries;
 import org.betonquest.betonquest.quest.registry.processor.VariableProcessor;
 import org.betonquest.betonquest.schedule.LastExecutionCache;
 import org.betonquest.betonquest.util.PlayerConverter;
