+++ src/main/java/org/betonquest/betonquest/api/feature/FeatureAPI.java
@@ -1,7 +1,7 @@
 package org.betonquest.betonquest.api.feature;
 
-import org.betonquest.betonquest.config.QuestCanceler;
 import org.betonquest.betonquest.conversation.ConversationData;
+import org.betonquest.betonquest.feature.QuestCanceler;
 import org.betonquest.betonquest.id.ConversationID;
 import org.betonquest.betonquest.id.QuestCancelerID;
 import org.betonquest.betonquest.quest.registry.QuestRegistry;
