+++ src/main/java/org/betonquest/betonquest/conversation/Conversation.java
@@ -17,12 +17,12 @@ import org.betonquest.betonquest.config.PluginMessage;
 import org.betonquest.betonquest.conversation.ConversationData.OptionType;
 import org.betonquest.betonquest.database.Saver.Record;
 import org.betonquest.betonquest.database.UpdateType;
+import org.betonquest.betonquest.feature.registry.ConversationIORegistry;
+import org.betonquest.betonquest.feature.registry.InterceptorRegistry;
 import org.betonquest.betonquest.id.ConditionID;
 import org.betonquest.betonquest.id.ConversationID;
 import org.betonquest.betonquest.id.EventID;
 import org.betonquest.betonquest.notify.Notify;
-import org.betonquest.betonquest.quest.registry.feature.ConversationIORegistry;
-import org.betonquest.betonquest.quest.registry.feature.InterceptorRegistry;
 import org.betonquest.betonquest.util.PlayerConverter;
 import org.betonquest.betonquest.util.Utils;
 import org.bukkit.Bukkit;
