+++ src/main/java/org/betonquest/betonquest/feature/registry/CoreFeatureFactories.java
@@ -1,4 +1,4 @@
-package org.betonquest.betonquest.quest.registry.feature;
+package org.betonquest.betonquest.feature.registry;
 
 import org.betonquest.betonquest.api.logger.BetonQuestLoggerFactory;
 import org.betonquest.betonquest.api.quest.QuestTypeAPI;
