+++ src/main/java/org/betonquest/betonquest/quest/registry/QuestRegistry.java
@@ -6,11 +6,11 @@ import org.betonquest.betonquest.api.logger.BetonQuestLogger;
 import org.betonquest.betonquest.api.logger.BetonQuestLoggerFactory;
 import org.betonquest.betonquest.bstats.InstructionMetricsSupplier;
 import org.betonquest.betonquest.config.PluginMessage;
+import org.betonquest.betonquest.feature.registry.FeatureRegistries;
+import org.betonquest.betonquest.feature.registry.processor.CancellerProcessor;
+import org.betonquest.betonquest.feature.registry.processor.ConversationProcessor;
 import org.betonquest.betonquest.id.ID;
-import org.betonquest.betonquest.quest.registry.feature.FeatureRegistries;
-import org.betonquest.betonquest.quest.registry.processor.CancellerProcessor;
 import org.betonquest.betonquest.quest.registry.processor.ConditionProcessor;
-import org.betonquest.betonquest.quest.registry.processor.ConversationProcessor;
 import org.betonquest.betonquest.quest.registry.processor.EventProcessor;
 import org.betonquest.betonquest.quest.registry.processor.ObjectiveProcessor;
 import org.betonquest.betonquest.quest.registry.processor.VariableProcessor;
